+++ src/interfaces/11_IElevator.sol
@@ -2,6 +2,6 @@
 pragma solidity ^0.8.23;
 
 interface IElevator {
-    function goTo(uint _floor) external;
-    function top() external view returns(bool);
+    function goTo(uint256 _floor) external;
+    function top() external view returns (bool);
 }
